consolidated recently viewed with requested widget
